+++ woocommerce/content-single-product.php
@@ -12,7 +12,8 @@ global $product;
 do_action('woocommerce_before_single_product');
 
 if (post_password_required()) {
-    echo get_the_password_form(); // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- Woo helper
+    // phpcs:ignore WordPress.Security.EscapeOutput.OutputNotEscaped -- Woo helper
+    echo get_the_password_form();
 
     return;
 }